Реализована отрисовка карточек в блоки
