+++ public/index.html
@@ -30,7 +30,7 @@
       <section class="films-list">
         <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
         <div class="films-list__container">
-          <!-- 2 карточки -->
+          <!-- Основной блок -->
         </div>
         <button class="films-list__show-more">Show more</button>
       </section>
