+++ src/main.js
@@ -1,5 +1,6 @@
-import renderFilter from './render-filter.js';
+import { filterClickHandler, renderFilter } from './render-filter.js';
 import renderCard from './render-card.js';
+import { getRandomNumber, clearContainer } from './utils.js';
 
 // Блок для вставки фильтра
 const filterContainer = document.querySelector(`.main-navigation`);
@@ -8,10 +9,7 @@ const filterContainer = document.querySelector(`.main-navigation`);
 const filmsContainers = Array.from(document.querySelectorAll(`.films-list__container`));
 
 // Стартовое кол-во карточек
-const startCardNumber = 7;
-
-// Список фильтров
-let cardsList = null;
+const cardsAmount = [7, 2, 2];
 
 // Список фильтров
 const filters = [
@@ -22,27 +20,15 @@ const filters = [
   `Stats`
 ];
 
-
-// Генерация случайного числа от min до max
-const getRandomNumber = (min, max) => {
-  return min + Math.floor(Math.random() * (max + 1 - min));
-};
-
-
-// Очистка блока от содержимого
-const clearContainer = (container) => {
-  container.innerHTML = ``;
-};
-
 // Обработчик клика по пункту фильтра
-const filterClickHandler = (evt) => {
-  evt.preventDefault();
-  clearContainer(filterContainer);
+// const filterClickHandler = (evt) => {
+//   evt.preventDefault();
+//   clearContainer(filterContainer);
 
-  const clickedFilter = evt.target;
+//   const clickedFilter = evt.target;
 
-  renderFilterList(cardNumber, filmsContainers);
-};
+//   renderFilterList(cardNumber, filmsContainers);
+// };
 
 // Обработчик клика по пункту фильтра
 const addFilterClickHandler = (element, amount) => {
@@ -52,7 +38,6 @@ const addFilterClickHandler = (element, amount) => {
   });
 };
 
-
 // Рендеринг фильтра
 const renderFilterList = (filtersArr, container) => {
   const fragment = document.createDocumentFragment();
@@ -60,7 +45,7 @@ const renderFilterList = (filtersArr, container) => {
   filtersArr.forEach((filterName, i) => {
     const first = (i === 0) ? true : null;
     const last = (i === filtersArr.length - 1) ? true : null;
-    const count = (first || last) ? 0 : getRandomNumber(0, 20);
+    const count = (first || last) ? 0 : getRandomNumber(1, 20);
 
     const filter = renderFilter(filterName, count, first, last);
 
@@ -68,10 +53,8 @@ const renderFilterList = (filtersArr, container) => {
   });
 };
 
-
 // Стартовый рендеринг фильтра
 renderFilterList(filters, filterContainer);
-
 
 // Отрисовка списка задач
 const renderCardList = (amount, container) => {
@@ -87,6 +70,7 @@ const renderCardList = (amount, container) => {
   container.appendChild(fragment);
 };
 
-
-// Запуск стартовой отрисовки карточек
-// renderTaskList(startTasksNumber, cardContainer);
+// Стартовая отрисовка карточек в основной блок
+renderCardList(cardsAmount[0], filmsContainers[0]);
+renderCardList(cardsAmount[1], filmsContainers[1]);
+renderCardList(cardsAmount[2], filmsContainers[2]);
